requirements.txt
torch
onnx
z3-solver
wget
onnx-simplifier
onnx-tf
onnxruntime-gpu
nvidia-pyindex
nvidia-tensorrt
pydot
gitpython
